Fixing warnings for 64 bit builds.
Former-commit-id: b839e3034c34c7f5e56e4e1c4db03305b2224744

+++ cocos2d/CCNode.m
@@ -1187,7 +1187,7 @@ CGAffineTransformMakeRigid(CGPoint translate, CGFloat radians)
 	return NO;
 }
 
--(CCTimer *) schedule:(SEL)selector interval:(CCTime)interval repeat: (uint) repeat delay:(CCTime) delay
+-(CCTimer *) schedule:(SEL)selector interval:(CCTime)interval repeat: (NSUInteger) repeat delay:(CCTime) delay
 {
 	NSAssert(selector != nil, @"Selector must be non-nil");
 	NSAssert(selector != @selector(update:) && selector != @selector(fixedUpdate:), @"The update: and fixedUpdate: methods are scheduled automatically.");
